Replace django.conf.urls.url with django.urls.re_path
This avoid a deprecation warning: django.conf.urls.url is an alias for
django.urls.re_path but will be removed in Django 4.0.

This method was only added in Django 2.0, so we have to check
whether we are using a version of Django greater than 2.0 before we
import it. Otherwise, we use the old `url` method.

Fixes #68